🐛 fix some none POSIX stuff and add confirmations

+++ setup.sh
@@ -1,10 +1,10 @@
 #!/bin/sh
 
 # User welcome message
-echo -e "\n####################################################################"
+printf '\n####################################################################\n'
 echo '# 👋 Welcome, this is the setup script for the tpn CLI tool.'
-echo -e "# Note: this script will ask for your password once or multiple times."
-echo -e "####################################################################\n\n"
+printf '# Note: this script will ask for your password once or multiple times.\n'
+printf '####################################################################\n\n'
 
 # Set the binary folder
 BINARY_FOLDER="/usr/local/bin"
@@ -14,16 +14,24 @@ BINARY_FILE="$BINARY_FOLDER/tpn"
 REPO_URL="https://raw.githubusercontent.com/taofu-labs/tpn-cli"
 FILE_URL="$REPO_URL/main/tpn.sh"
 
+# Ask user Y/n if they want to proceed
+read -p "Do you want to proceed with the installation? (Y/n): " response
+response=$(echo "$response" | tr '[:upper:]' '[:lower:]')  # Convert to lowercase
+if [ "$response" = "n" ] || [ "$response" = "no" ]; then
+    echo "Installation aborted by user."
+    exit 0
+fi
+
 # Ask for sudo once, in most systems this will cache the permissions for a bit
 sudo echo "Starting TPN installation"
-echo -e "[ 1 ] Superuser permissions acquired."
+printf '[ 1 ] Superuser permissions acquired.\n'
 
 # Make sure binfolder exists
-echo -e "[ 2 ] Ensuring the binary folder exists at $BINARY_FOLDER"
+printf '[ 2 ] Ensuring the binary folder exists at %s\n' "$BINARY_FOLDER"
 sudo mkdir -p "$BINARY_FOLDER"
 
 # Download the file to tempfolder and move it to the binary folder
-echo -e "[ 3 ] Downloading the TPN CLI tool from Github"
+printf '[ 3 ] Downloading the TPN CLI tool from Github\n'
 sudo rm -f "$BINARY_FILE"
 TEMP_DIR=$(mktemp -d -t tpn_download_XXXXXX)
 curl -sSL "$FILE_URL" -o "$TEMP_DIR/tpn.sh"
@@ -35,7 +43,7 @@ OWNER=$(stat -f '%Su' "$BINARY_FILE")
 
 # If user is not owner, change ownership
 if [ "$OWNER" != "$USER" ]; then
-    echo -e "[ ! ] Changing ownership of $BINARY_FILE to $USER"
+    printf '[ ! ] Changing ownership of %s to %s\n' "$BINARY_FILE" "$USER"
     sudo chown "$USER" "$BINARY_FILE"
 fi
 
@@ -44,18 +52,18 @@ PERMS=$(stat -f '%Mp%Lp' "$BINARY_FILE")
 
 # If permissions not 755, change them
 if [ "$PERMS" != "0755" ]; then
-    echo -e "[ ! ] Changing permissions of $BINARY_FILE to 755"
+    printf '[ ! ] Changing permissions of %s to 755\n' "$BINARY_FILE"
     sudo chmod 755 "$BINARY_FILE"
 fi
 
 # If not yet executable, make it executable
 if [ ! -x "$BINARY_FILE" ]; then
-    echo -e "[ ! ] Making $BINARY_FILE executable"
+    printf '[ ! ] Making %s executable\n' "$BINARY_FILE"
     sudo chmod +x "$BINARY_FILE"
 fi
 
 # Final message
-echo -e "\n####################################################################"
+printf '\n####################################################################\n'
 echo "# 🎉 TPN CLI tool has been successfully installed at $BINARY_FILE"
 echo "# You can now use it by typing 'tpn' in your terminal."
-echo -e "####################################################################\n"
+printf '####################################################################\n'
